change the type of time in sleep() into unsigned long #18

+++ src/Zackernel.cpp
@@ -47,7 +47,7 @@ void dispatch() {
   }  
 }
 
-void sleep(int time, vl::Func<void(void)> func) {
+void sleep(unsigned long time, vl::Func<void(void)> func) {
   Schedule::add(time, func);
   dispatch();
 }
